requirements.txt
# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-upstage>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10

# Vector DB
chromadb>=0.4.0

# OpenAI
openai>=1.0.0

# Environment
python-dotenv>=1.0.0

# UI
streamlit>=1.30.0
